# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bank Statement Analyzer Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Budgets
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List budgets with actual vs planned spending
      responses:
        "200":
          description: Budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"

    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    patch:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"

    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Goals
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List financial goals
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"

    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a financial goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Analytics (wow endpoints)
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List generated reports
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"

    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # User Profile
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth
...
    Category:
      type: object
      required: [id, name, icon, color, type, isDefault]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
        isDefault:
          type: boolean
        transactionCount:
          type: ["integer", "null"]
    CategoryInput:
      type: object
      required: [name, icon, color, type]
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string

    # Budgets
    Budget:
      type: object
      required: [id, categoryId, categoryName, period, amount, spent, remaining]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: string
        period:
          type: string
          enum: [monthly, yearly]
        amount:
          type: number
        spent:
          type: number
        remaining:
          type: number
        percentUsed:
          type: number
        month:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
    BudgetInput:
      type: object
      required: [categoryId, period, amount]
      properties:
        categoryId:
          type: integer
        period:
          type: string
          enum: [monthly, yearly]
        amount:
          type: number
        month:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
    BudgetUpdate:
      type: object
      properties:
        amount:
          type: number
        period:
          type: ["string", "null"]
          enum: [monthly, yearly, null]

    # Goals
    Goal:
      type: object
      required: [id, name, type, targetAmount, currentAmount, deadline, progressPercent]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [emergency_fund, vehicle, vacation, house, custom]
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: ["string", "null"]
        progressPercent:
          type: number
        estimatedCompletion:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    GoalInput:
      type: object
      required: [name, type, targetAmount, currentAmount]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [emergency_fund, vehicle, vacation, house, custom]
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    GoalUpdate:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: ["number", "null"]
        currentAmount:
          type: ["number", "null"]
        deadline:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # Analytics
...
    Report:
      type: object
      required: [id, type, title, period, status, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [monthly, annual, tax, business, executive]
        title:
          type: string
        period:
          type: string
        status:
          type: string
          enum: [generating, ready, error]
        summary:
          type: ["string", "null"]
        data:
          type: ["object", "null"]
        createdAt:
          type: string
    ReportInput:
      type: object
      required: [type, period]
      properties:
        type:
          type: string
          enum: [monthly, annual, tax, business, executive]
        period:
          type: string
        title:
          type: ["string", "null"]

    # Admin
...